добавлены компоненты и лэйаут
