--- pages/404.js
@@ -1,8 +1,0 @@
-const Error = () => (
-    <>
-        <h1>404</h1>
-        <h2>Something is going wrong...</h2>
-    </>
-);
-
-export default Error
