--- pages/_app.js
@@ -1,7 +1,0 @@
-import '../styles/globals.css'
-
-const App = ({ Component, pageProps }) => (
-  <Component {...pageProps} />
-);
-
-export default App
