--- pages/contacts/contact.js
@@ -1,5 +1,0 @@
-const Contact = () => (
-    <h1>Jack</h1>
-);
-
-export default Contact
