--- pages/contacts/index.js
@@ -1,7 +1,0 @@
-const Contacts = () => (
-    <>
-        <h1>Contacts list:</h1>
-    </>
-);
-
-export default Contacts
