--- pages/index.js
@@ -1,5 +1,0 @@
-const Home = () => (
-  <h1>Hello world!</h1>
-);
-
-export default Home
