+++ src/components/Layout/Layout.js
@@ -1,0 +1,12 @@
+import Header from "../ui/Header/Header";
+import Footer from "../ui/Footer/Footer";
+
+const Layout = ({ children }) => (
+    <>
+        <Header />
+        {children}
+        <Footer />
+    </>
+)
+
+export default Layout
