+++ src/components/ui/Footer/Footer.js
@@ -1,0 +1,9 @@
+import Heading from "../Heading/Heading";
+
+const Footer = () => (
+    <footer>
+        <Heading tag="h3" text="Footer" />
+    </footer>
+);
+
+export default Footer
