+++ src/components/ui/Header/Header.js
@@ -1,0 +1,9 @@
+import Heading from "../Heading/Heading";
+
+const Header = () => (
+    <header>
+        <Heading tag="h3" text="Header" />
+    </header>
+);
+
+export default Header
