+++ src/components/ui/Heading/Heading.js
@@ -1,0 +1,7 @@
+const Heading = ({ tag, text }) => {
+    const Tag = tag || 'h1';
+
+    return <Tag>{text}</Tag>
+};
+
+export default Heading
