+++ src/pages/404.js
@@ -1,0 +1,10 @@
+import Heading from "../components/ui/Heading/Heading";
+
+const Error = () => (
+    <>
+        <Heading text="404" />
+        <Heading tag="h2" text="Something is going wrong..." />
+    </>
+);
+
+export default Error
