+++ src/pages/contacts/contact.js
@@ -1,0 +1,7 @@
+import Heading from "../../components/ui/Heading/Heading";
+
+const Contact = () => (
+    <Heading text="Jack" />
+);
+
+export default Contact
