+++ src/pages/contacts/index.js
@@ -1,0 +1,7 @@
+import Heading from "../../components/ui/Heading/Heading";
+
+const Contacts = () => (
+    <Heading text="Contacts list:" />
+);
+
+export default Contacts
