+++ src/pages/index.js
@@ -1,0 +1,7 @@
+import Heading from "../components/ui/Heading/Heading";
+
+const Home = () => (
+  <Heading text="Hello world!" />
+);
+
+export default Home
